[refactor] add shared datalink for all firmware
[refactor] add shared datalink for all firmware

[refactor] add shared datalink for all firmware

[refactor] add shared datalink for all firmware
