+++ conf/firmwares/subsystems/fixedwing/fdm_jsbsim.makefile
@@ -67,5 +67,5 @@ nps.srcs += $(NPSDIR)/nps_main.c                 \
 nps.srcs += math/pprz_geodetic_wmm2015.c
 
 include $(CFG_SHARED)/telemetry_transparent_udp.makefile
-nps.srcs += $(SRC_FIRMWARE)/datalink.c
+nps.srcs += subsystems/datalink/datalink.c $(SRC_FIRMWARE)/fixedwing_datalink.c
 nps.srcs += $(SRC_FIRMWARE)/ap_downlink.c $(SRC_FIRMWARE)/fbw_downlink.c
