+++ conf/firmwares/subsystems/fixedwing/telemetry_bluegiga.makefile
@@ -5,7 +5,7 @@ ifeq ($(TARGET),ap)
 include $(CFG_SHARED)/telemetry_bluegiga.makefile
 endif
 
-ap.srcs += $(SRC_FIRMWARE)/datalink.c $(SRC_FIRMWARE)/ap_downlink.c
+ap.srcs += $(SRC_FIRMWARE)/fixedwing_datalink.c $(SRC_FIRMWARE)/ap_downlink.c
 
 # avoid fbw_telemetry_mode error
 ap.srcs += $(SRC_FIRMWARE)/fbw_downlink.c
