+++ conf/firmwares/subsystems/fixedwing/telemetry_transparent_usb.makefile
@@ -6,7 +6,7 @@ ifeq ($(TARGET), ap)
 include $(CFG_SHARED)/telemetry_transparent_usb.makefile
 endif
 
-ap.srcs += $(SRC_FIRMWARE)/datalink.c $(SRC_FIRMWARE)/ap_downlink.c $(SRC_FIRMWARE)/fbw_downlink.c
+ap.srcs += $(SRC_FIRMWARE)/fixedwing_datalink.c $(SRC_FIRMWARE)/ap_downlink.c $(SRC_FIRMWARE)/fbw_downlink.c
 # avoid fbw_telemetry_mode error
 ap.srcs += $(SRC_FIRMWARE)/fbw_downlink.c
