+++ conf/firmwares/subsystems/rotorcraft/telemetry_bluegiga.makefile
@@ -6,4 +6,4 @@ ifeq ($(TARGET), ap)
 include $(CFG_SHARED)/telemetry_bluegiga.makefile
 endif
 
-ap.srcs += $(SRC_FIRMWARE)/datalink.c $(SRC_FIRMWARE)/rotorcraft_telemetry.c
+ap.srcs += $(SRC_FIRMWARE)/rotorcraft_datalink.c $(SRC_FIRMWARE)/rotorcraft_telemetry.c
