+++ conf/firmwares/subsystems/rotorcraft/telemetry_superbitrf.makefile
@@ -8,4 +8,4 @@ include $(CFG_SHARED)/telemetry_superbitrf.makefile
 endif
 
 # add rotorcraft specific files
-ap.srcs += $(SRC_FIRMWARE)/datalink.c $(SRC_FIRMWARE)/rotorcraft_telemetry.c
+ap.srcs += $(SRC_FIRMWARE)/rotorcraft_datalink.c $(SRC_FIRMWARE)/rotorcraft_telemetry.c
