+++ conf/firmwares/subsystems/rotorcraft/telemetry_transparent_usb.makefile
@@ -5,4 +5,4 @@ ifeq ($(TARGET), ap)
 include $(CFG_SHARED)/telemetry_transparent_usb.makefile
 endif
 
-ap.srcs += $(SRC_FIRMWARE)/datalink.c $(SRC_FIRMWARE)/rotorcraft_telemetry.c
+ap.srcs += $(SRC_FIRMWARE)/rotorcraft_datalink.c $(SRC_FIRMWARE)/rotorcraft_telemetry.c
