+++ conf/firmwares/subsystems/rotorcraft/telemetry_xbee_api.makefile
@@ -10,4 +10,4 @@ ifeq ($(TARGET), ap)
 include $(CFG_SHARED)/telemetry_xbee_api.makefile
 endif
 
-ap.srcs += $(SRC_FIRMWARE)/datalink.c $(SRC_FIRMWARE)/rotorcraft_telemetry.c
+ap.srcs += $(SRC_FIRMWARE)/rotorcraft_datalink.c $(SRC_FIRMWARE)/rotorcraft_telemetry.c
