+++ conf/firmwares/subsystems/shared/telemetry_bluegiga.makefile
@@ -31,5 +31,5 @@ ifneq ($(MODEM_LED),none)
 ap.CFLAGS += -DMODEM_LED=$(MODEM_LED)
 endif
 
-ap.srcs += $(SRC_SUBSYSTEMS)/datalink/downlink.c $(SRC_SUBSYSTEMS)/datalink/bluegiga.c
+ap.srcs += $(SRC_SUBSYSTEMS)/datalink/downlink.c subsystems/datalink/datalink.c $(SRC_SUBSYSTEMS)/datalink/bluegiga.c
 ap.srcs += $(PAPARAZZI_HOME)/var/share/pprzlink/src/pprz_transport.c $(SRC_SUBSYSTEMS)/datalink/telemetry.c
