+++ conf/firmwares/subsystems/shared/telemetry_transparent_udp.makefile
@@ -19,5 +19,5 @@ TELEM_CFLAGS += -DDOWNLINK_TRANSPORT=pprz_tp -DDATALINK=PPRZ
 
 
 $(TARGET).CFLAGS += $(MODEM_CFLAGS) $(TELEM_CFLAGS)
-$(TARGET).srcs += subsystems/datalink/downlink.c $(PAPARAZZI_HOME)/var/share/pprzlink/src/pprz_transport.c subsystems/datalink/telemetry.c
+$(TARGET).srcs += subsystems/datalink/downlink.c subsystems/datalink/datalink.c $(PAPARAZZI_HOME)/var/share/pprzlink/src/pprz_transport.c subsystems/datalink/telemetry.c
